requirements.txt
python-debian
requests
tqdm